WIP: moving cluster store.
